refactor: split main and lib
